+++ python/sgl_jax/srt/mem_cache/allocator.py
@@ -443,11 +443,11 @@ class SWATokenToKVPoolAllocator(BaseTokenToKVPoolAllocator):
             )
         if dp_size == 1:
             self.full_to_swa_index_mapping = np.zeros(
-                self.full_attn_allocator.size_per_rank + 1, dtype=np.int64
+                self.full_attn_allocator.size_per_rank + 1, dtype=np.int32
             )
         else:
             self.full_to_swa_index_mapping = [
-                np.zeros(self.full_attn_allocator.size_per_rank + 1, dtype=np.int64)
+                np.zeros(self.full_attn_allocator.size_per_rank + 1, dtype=np.int32)
                 for _ in range(dp_size)
             ]
         self.clear()
